feat: 날몹 구현

+++ CityScape/Assets/Scripts/Monster/Monster.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Monster : MonoBehaviour
 {
-    public float id;
+    public int id;
     public float prefabID;
 
     public float health;
